+++ UnitTest++/ArgumentsReader.h
@@ -11,9 +11,9 @@ namespace UnitTest {
 		ArgumentsReader(int argc, char**argv);
 		virtual ~ArgumentsReader();
 
-		bool findArgumentListIndex(const string & argumentName, int & outFrom, int & outCount) const;
+		bool findArgumentListIndex(const string & argumentName, size_t & outFrom, size_t & outCount) const;
 		vector<string> extractValues(const string & argumentName) const;
-		string getArgument(int index) const;
+		string getArgument(size_t index) const;
 
 	private:
 		vector<string> _arguments;
